Moved check_sufficient_finality_signatures to utils
 - Simplified arithmetic
