+++ node/src/components/linear_chain_sync.rs
@@ -42,6 +42,7 @@ use datasize::DataSize;
 use prometheus::Registry;
 use tracing::{error, info, trace, warn};
 
+use casper_execution_engine::{shared::stored_value::StoredValue, storage::trie::Trie};
 use casper_types::{EraId, Key, ProtocolVersion};
 
 use self::event::{BlockByHashResult, DeploysResult};
@@ -68,12 +69,11 @@ use crate::{
     },
     NodeRng,
 };
-use casper_execution_engine::{shared::stored_value::StoredValue, storage::trie::Trie};
+
 pub(crate) use error::FinalitySignatureError;
 use event::BlockByHeightResult;
 pub use event::Event;
 pub use metrics::LinearChainSyncMetrics;
-pub(crate) use operations::check_sufficient_finality_signatures;
 pub use peers::PeersState;
 pub use state::State;
 pub use traits::ReactorEventT;
